+++ aiosend/webhook/__init__.py
@@ -1,14 +1,17 @@
 from .aiohttp_manager import AiohttpManager
 from .base import _APP, WebhookHandler, WebhookManager
-from .fastapi_manager import FastAPIManager
 from .flask_manager import FlaskManager
 from .router import WebhookRouter
+from .starlette_manager import StarletteManager
+
+# from .fastapi_manager import FastAPIManager
 
 __all__ = (
     "_APP",
     "AiohttpManager",
-    "FastAPIManager",
     "FlaskManager",
+    "StarletteManager",
+    # "FastAPIManager",
     "WebhookHandler",
     "WebhookManager",
     "WebhookRouter",
