+++ examples/webhook_fastapi.py
@@ -11,13 +11,16 @@ cp = CryptoPay(
     webhook_manager=FastAPIManager(app, "/handler"),
 )
 
+
 @cp.invoice_paid()
 async def handler(invoice: Invoice) -> None:
     print(f"Received {invoice.amount} {invoice.asset}")
 
+
 async def main() -> None:
     invoice = await cp.create_invoice(1, "USDT")
     print("invoice link:", invoice.bot_invoice_url)
+
 
 if __name__ == "__main__":
     asyncio.run(main())
